requirements.txt
flask>=2.2
flask_sqlalchemy>=3.0
flask_migrate>=4.0
python-dotenv>=1.0
psycopg2-binary>=2.9
flask-cors>=3.0.1
PyJWT>=2.0